+++ pineappl/src/fk_table.rs
@@ -1,8 +1,8 @@
 //! Provides the [`FkTable`] type.
 
 use super::boc::Order;
+use super::convolutions::LumiCache;
 use super::grid::{Convolution, Grid, GridError};
-use super::lumi::LumiCache;
 use super::subgrid::Subgrid;
 use float_cmp::approx_eq;
 use ndarray::Array4;